Add comment

+++ frontend/src/components/ImageDetail/index.js
@@ -2,9 +2,10 @@ import React, { useEffect, useState } from "react";
 import { useParams, useHistory } from "react-router-dom"
 import { useSelector, useDispatch } from "react-redux";
 import { thunkDeleteImage, thunkGetAllImages, thunkGetOneImage, thunkUpdateImage } from "../../store/image";
-import { ImageDeletePage } from "../ImageDeletePage";
+import { thunkGetComments, thunkDeleteComment } from "../../store/comment";
 import ImageUpdatePage from "../ImageEditPage";
-
+import CommentCreate from "../CommentCreate/commentCreate";
+// import { thunkCreateComment } from "../../store/comment";
 
 const ImageDetail = () => {
   const dispatch = useDispatch();
@@ -12,35 +13,61 @@ const ImageDetail = () => {
   const user = useSelector(state => state.session.user)
   const history = useHistory();
   const [imageUrl, setImageUrl] = useState("");
+  const [selectComment, setSelectComment] = useState("");
   const [showEditImageForm, setShowEditImageForm] = useState(false);
+  const [isLoaded, setIsLoaded] = useState(false);
   // const [updateImage, setUpdateImage] = useState(false);
 
 
-  // console.log("+++++++++++component imageDetail+++++++: ", user);
   const image = useSelector(state => state?.images[imageId]);
+  // console.log("+++++++++++component imageDetail+++++++: ", image);
 
-  // const images = useSelector(state => state.images);
-  // const imageArr = Object.values(images)
-  // console.log("+++++++++++component imageDetail+++++++: ", imageArr)
-  // const image = imageArr?.imageId
+  // const comment = useSelector(state => state?.Comments)
+  // console.log("+++++++++++component imageDetail+++++++: ", comment);
+
+  // const comment = image.Comments
+  // console.log("+++++++++++component imageDetail+++++++: ", comment);
+
 
   useEffect(() => {
     // dispatch(thunkGetOneImage(imageId));
-    dispatch(thunkGetAllImages());
-  }, [dispatch]);
+    // dispatch(thunkGetComments())
+    dispatch(thunkGetAllImages())
+      .then(() => setIsLoaded(true));
+  }, [dispatch, isLoaded]);
 
   const deleteImage = async e => {
     e.preventDefault();
-    await dispatch(thunkDeleteImage(imageId));
-    history.push('/images')
+    const confirmed = window.confirm("Are you sure you want to delete this image?")
+    if (confirmed) {
+      await dispatch(thunkDeleteImage(imageId));
+      history.push('/images')
+    }
+
   }
 
-  // const editImage = async e => {
-  //   await dispatch(thunkUpdateImage(imageId));
+  // const deleteComment = async e => {
+  //   e.preventDefault();
+  //   const comment = image.Comments
+  //   // console.log("+++++++++++component imageDetail+++++++: ", comment);
+
+  //   let sortedComment = {};
+  //   comment.map(element => sortedComment[element.id] = element)
+  //   console.log(sortedComment)
+
+  //   // console.log("&&&&&&&&&&&&& Delete Comment&&&&&&&&&: ", comment.id)
+  //   // await dispatch(thunkDeleteComment());
   //   history.push(`/images/${imageId}`)
   // }
 
-
+  const deleteComment = async (commentId) => {
+    // console.log("hi",commentId)
+    const confirmed = window.confirm("Are you sure you want to delete this comment?")
+    if (confirmed) {
+      return await dispatch(thunkDeleteComment(commentId))
+        .then(setIsLoaded(!isLoaded))
+    }
+  }
   // console.log("+++++++++++component imageDetail+++++++: ", image);
 
   const updateImage = (e) => {
@@ -51,10 +78,8 @@ const ImageDetail = () => {
       <>
         <div id="image-container">
           <h1>Image Detail</h1>
-          <img src={image?.imageUrl} alt="image"/>
+          <img src={image?.imageUrl} key={imageId} alt="image"/>
             {user?.id === image?.userId && <button onClick={updateImage}>Edit Image</button>}
-            {/* {user?.id === image?.userId && <button onClick={() => setShowEditImageForm(true)}>Edit Image</button>} */}
-            {/* {user?.id === image?.userId && <button>Edit Image</button>} */}
 
             {
               showEditImageForm &&
@@ -67,27 +92,30 @@ const ImageDetail = () => {
               </div>
             }
 
-            {/* <form hideForm={() => setShowEditImageForm(true)} id="image-edit-form">
-              <label>
-                Image Url
-                <input
-                  type='text'
-                  value={image.imageUrl}
-                  onChange={(e) => setImageUrl(e.target.value)}
-                  required
-                />
-              </label>
-              <button type="submit">Confirm</button>
-            </form> */}
-
             {user?.id === image?.userId && <button onClick={deleteImage}>Delete Image</button>}
+
           <div id="comment-component">
-            <div>
-              {
-                image?.Comments?.map(comment => <div key={comment.id}>{comment.message}</div>)
-              }
-            </div>
+            {
+              image?.Comments?.map(comment =>
+                <div key={comment.id}>
+                  {comment.message}
+                  {user?.id === comment.userId && <button onClick={() => deleteComment(comment.id)}>Delete Comment</button>}
+                </div>)
+            }
           </div>
+
+          {
+            user?
+              <div id="create-comment-component">
+                <CommentCreate image={image}/>
+              </div>
+              :
+              <div id="redirect-to-login-page">
+                <p>Please log in to comment!</p>
+              </div>
+          }
+
+
         </div>
       </>
   );
